8/31 초기취향페이지 완료! 아작스 완료! 꺅!

+++ mint_project_01/src/main/java/com/mint/project/dtos/TasteDto.java
@@ -4,35 +4,35 @@ public class TasteDto {
 
 	   private int tseq;
 	   private int tuseq;
-	   private long taction;
-	   private long tcomedy;
-	   private long tcrime;
-	   private long tdocumentary;
-	   private long tdrama;
-	   private long tfamily;
-	   private long tfantasy;
-	   private long tnoir;
-	   private long thorror;
-	   private long tmusical;
-	   private long tmystery;
-	   private long tromance;
-	   private long tsf;
-	   private long tsports;
-	   private long tthriller;
-	   private long twar;
-	   private long tcar;
-	   private long trabbit;
-	   private long tanimation;
-	   private long tchildren;
-	   private long thistory;
-	   private long troco;
-	   private long tdisaster;
-	   private long twuxia;
-	   private long twestern;
-	   private long thotguy;
-	   private long thighteen;
-	   private long tdomestic;
-	   private long tforeign;
+	   private int taction;
+	   private int tcomedy;
+	   private int tcrime;
+	   private int tdocumentary;
+	   private int tdrama;
+	   private int tfamily;
+	   private int tfantasy;
+	   private int tnoir;
+	   private int thorror;
+	   private int tmusical;
+	   private int tmystery;
+	   private int tromance;
+	   private int tsf;
+	   private int tsports;
+	   private int tthriller;
+	   private int twar;
+	   private int tcar;
+	   private int trabbit;
+	   private int tanimation;
+	   private int tchildren;
+	   private int thistory;
+	   private int troco;
+	   private int tdisaster;
+	   private int twuxia;
+	   private int twestern;
+	   private int thotguy;
+	   private int thighteen;
+	   private int tdomestic;
+	   private int tforeign;
 	   private String tstatus;
 	   
 	   public TasteDto() {
@@ -41,11 +41,11 @@ public class TasteDto {
 	   }
 	   
 
-	   public TasteDto(int tseq, int tuseq, long taction, long tcomedy, long tcrime, long tdocumentary,
-	         long tdrama, long tfamily, long tfantasy, long tnoir, long thorror, long tmusical,
-	         long tmystery, long tromance, long tsf, long tsports, long tthriller, long twar, long tcar,
-	         long trabbit, long tanimation, long tchildren, long thistory, long troco, long tdisaster,
-	         long twuxia, long twestern, long thotguy, long thighteen, long tdomestic, long tforeign, String tstatus) {
+	   public TasteDto(int tseq, int tuseq, int taction, int tcomedy, int tcrime, int tdocumentary,
+	         int tdrama, int tfamily, int tfantasy, int tnoir, int thorror, int tmusical,
+	         int tmystery, int tromance, int tsf, int tsports, int tthriller, int twar, int tcar,
+	         int trabbit, int tanimation, int tchildren, int thistory, int troco, int tdisaster,
+	         int twuxia, int twestern, int thotguy, int thighteen, int tdomestic, int tforeign, String tstatus) {
 	      super();
 	      this.tseq = tseq;
 	      this.tuseq = tuseq;
@@ -95,178 +95,178 @@ public class TasteDto {
 	   public void setTuseq(int tuseq) {
 	      this.tuseq = tuseq;
 	   }
-	   public long getTaction() {
+	   public int getTaction() {
 	      return taction;
 	   }
-	   public void setTaction(long taction) {
+	   public void setTaction(int taction) {
 	      this.taction = taction;
 	   }
-	   public long getTcomedy() {
+	   public int getTcomedy() {
 	      return tcomedy;
 	   }
-	   public void setTcomedy(long tcomedy) {
+	   public void setTcomedy(int tcomedy) {
 	      this.tcomedy = tcomedy;
 	   }
-	   public long getTcrime() {
+	   public int getTcrime() {
 	      return tcrime;
 	   }
-	   public void setTcrime(long tcrime) {
+	   public void setTcrime(int tcrime) {
 	      this.tcrime = tcrime;
 	   }
-	   public long getTdocumentary() {
+	   public int getTdocumentary() {
 	      return tdocumentary;
 	   }
-	   public void setTdocumentary(long tdocumentary) {
+	   public void setTdocumentary(int tdocumentary) {
 	      this.tdocumentary = tdocumentary;
 	   }
-	   public long getTdrama() {
+	   public int getTdrama() {
 	      return tdrama;
 	   }
-	   public void setTdrama(long tdrama) {
+	   public void setTdrama(int tdrama) {
 	      this.tdrama = tdrama;
 	   }
-	   public long getTfamily() {
+	   public int getTfamily() {
 	      return tfamily;
 	   }
-	   public void setTfamily(long tfamily) {
+	   public void setTfamily(int tfamily) {
 	      this.tfamily = tfamily;
 	   }
-	   public long getTfantasy() {
+	   public int getTfantasy() {
 	      return tfantasy;
 	   }
-	   public void setTfantasy(long tfantasy) {
+	   public void setTfantasy(int tfantasy) {
 	      this.tfantasy = tfantasy;
 	   }
-	   public long getTnoir() {
+	   public int getTnoir() {
 	      return tnoir;
 	   }
-	   public void setTnoir(long tnoir) {
+	   public void setTnoir(int tnoir) {
 	      this.tnoir = tnoir;
 	   }
-	   public long getThorror() {
+	   public int getThorror() {
 	      return thorror;
 	   }
-	   public void setThorror(long thorror) {
+	   public void setThorror(int thorror) {
 	      this.thorror = thorror;
 	   }
-	   public long getTmusical() {
+	   public int getTmusical() {
 	      return tmusical;
 	   }
-	   public void setTmusical(long tmusical) {
+	   public void setTmusical(int tmusical) {
 	      this.tmusical = tmusical;
 	   }
-	   public long getTmystery() {
+	   public int getTmystery() {
 	      return tmystery;
 	   }
-	   public void setTmystery(long tmystery) {
+	   public void setTmystery(int tmystery) {
 	      this.tmystery = tmystery;
 	   }
-	   public long getTromance() {
+	   public int getTromance() {
 	      return tromance;
 	   }
-	   public void setTromance(long tromance) {
+	   public void setTromance(int tromance) {
 	      this.tromance = tromance;
 	   }
-	   public long getTsf() {
+	   public int getTsf() {
 	      return tsf;
 	   }
-	   public void setTsf(long tsf) {
+	   public void setTsf(int tsf) {
 	      this.tsf = tsf;
 	   }
-	   public long getTsports() {
+	   public int getTsports() {
 	      return tsports;
 	   }
-	   public void setTsports(long tsports) {
+	   public void setTsports(int tsports) {
 	      this.tsports = tsports;
 	   }
-	   public long getTthriller() {
+	   public int getTthriller() {
 	      return tthriller;
 	   }
-	   public void setTthriller(long tthriller) {
+	   public void setTthriller(int tthriller) {
 	      this.tthriller = tthriller;
 	   }
-	   public long getTwar() {
+	   public int getTwar() {
 	      return twar;
 	   }
-	   public void setTwar(long twar) {
+	   public void setTwar(int twar) {
 	      this.twar = twar;
 	   }
-	   public long getTcar() {
+	   public int getTcar() {
 	      return tcar;
 	   }
-	   public void setTcar(long tcar) {
+	   public void setTcar(int tcar) {
 	      this.tcar = tcar;
 	   }
-	   public long getTrabbit() {
+	   public int getTrabbit() {
 	      return trabbit;
 	   }
-	   public void setTrabbit(long trabbit) {
+	   public void setTrabbit(int trabbit) {
 	      this.trabbit = trabbit;
 	   }
-	   public long getTanimation() {
+	   public int getTanimation() {
 	      return tanimation;
 	   }
-	   public void setTanimation(long tanimation) {
+	   public void setTanimation(int tanimation) {
 	      this.tanimation = tanimation;
 	   }
-	   public long getTchildren() {
+	   public int getTchildren() {
 	      return tchildren;
 	   }
-	   public void setTchildren(long tchildren) {
+	   public void setTchildren(int tchildren) {
 	      this.tchildren = tchildren;
 	   }
-	   public long getThistory() {
+	   public int getThistory() {
 	      return thistory;
 	   }
-	   public void setThistory(long thistory) {
+	   public void setThistory(int thistory) {
 	      this.thistory = thistory;
 	   }
-	   public long getTroco() {
+	   public int getTroco() {
 	      return troco;
 	   }
-	   public void setTroco(long troco) {
+	   public void setTroco(int troco) {
 	      this.troco = troco;
 	   }
-	   public long getTdisaster() {
+	   public int getTdisaster() {
 	      return tdisaster;
 	   }
-	   public void setTdisaster(long tdisaster) {
+	   public void setTdisaster(int tdisaster) {
 	      this.tdisaster = tdisaster;
 	   }
-	   public long getTwuxia() {
+	   public int getTwuxia() {
 	      return twuxia;
 	   }
-	   public void setTwuxia(long twuxia) {
+	   public void setTwuxia(int twuxia) {
 	      this.twuxia = twuxia;
 	   }
-	   public long getTwestern() {
+	   public int getTwestern() {
 	      return twestern;
 	   }
-	   public void setTwestern(long twestern) {
+	   public void setTwestern(int twestern) {
 	      this.twestern = twestern;
 	   }
-	   public long getThotguy() {
+	   public int getThotguy() {
 	      return thotguy;
 	   }
-	   public void setThotguy(long thotguy) {
+	   public void setThotguy(int thotguy) {
 	      this.thotguy = thotguy;
 	   }
-	   public long getThighteen() {
+	   public int getThighteen() {
 	      return thighteen;
 	   }
-	   public void setThighteen(long thighteen) {
+	   public void setThighteen(int thighteen) {
 	      this.thighteen = thighteen;
 	   }
-	   public long getTdomestic() {
+	   public int getTdomestic() {
 	      return tdomestic;
 	   }
-	   public void setTdomestic(long tdomestic) {
+	   public void setTdomestic(int tdomestic) {
 	      this.tdomestic = tdomestic;
 	   }
-	   public long getTforeign() {
+	   public int getTforeign() {
 	      return tforeign;
 	   }
-	   public void setTforeign(long tforeign) {
+	   public void setTforeign(int tforeign) {
 	      this.tforeign = tforeign;
 	   }
 	   public String getTstatus() {
